requirements.txt
bcrypt==3.1.7
dateparser==0.7.4
Flask==1.1.1
Flask-API==2.0
Flask-JWT-Extended==3.24.1
gunicorn==20.0.4
httplib2==0.17.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
peewee==3.13.2
PyJWT==1.7.1
mysqlclient==1.4.6
python-dateutil==2.8.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
urllib3==1.25.8
Werkzeug==1.0.1